+++ astropy/units/decorators.py
@@ -3,11 +3,15 @@
 
 __all__ = ['quantity_input']
 
+from numbers import Number
 from collections.abc import Sequence
 import inspect
-from astropy.utils.decorators import wraps
 
-from .core import Unit, UnitBase, UnitsError, add_enabled_equivalencies
+import numpy as np
+
+from astropy.utils.decorators import wraps
+from .core import (Unit, UnitBase, UnitsError, add_enabled_equivalencies,
+                   dimensionless_unscaled)
 from .physical import _unit_physical_mapping
 
 
@@ -29,8 +33,7 @@ def _get_allowed_units(targets):
                 physical_type_id = _unit_physical_mapping[target]
 
             except KeyError:  # Function argument target is invalid
-                raise ValueError("Invalid unit or physical type '{}'."
-                                 .format(target))
+                raise ValueError(f"Invalid unit or physical type '{target}'.")
 
             # get unit directly from physical type id
             target_unit = Unit._from_physical_type_id(physical_type_id)
@@ -40,7 +43,8 @@ def _get_allowed_units(targets):
     return allowed_units
 
 
-def _validate_arg_value(param_name, func_name, arg, targets, equivalencies):
+def _validate_arg_value(param_name, func_name, arg, targets, equivalencies,
+                        strict_dimensionless=False):
     """
     Validates the object passed in to the wrapped function, ``arg``, with target
     unit or physical type, ``target``.
@@ -50,6 +54,16 @@ def _validate_arg_value(param_name, func_name, arg, targets, equivalencies):
         return
 
     allowed_units = _get_allowed_units(targets)
+
+    # If dimensionless is an allowed unit, allow numbers or numpy arrays with
+    #  numeric dtypes:
+    if dimensionless_unscaled in allowed_units and not strict_dimensionless:
+        if isinstance(arg, Number):
+            return
+
+        elif (isinstance(arg, np.ndarray)
+              and np.issubdtype(arg.dtype, np.number)):
+            return
 
     for allowed_unit in allowed_units:
         try:
@@ -61,25 +75,23 @@ def _validate_arg_value(param_name, func_name, arg, targets, equivalencies):
 
         except AttributeError:  # Either there is no .unit or no .is_equivalent
             if hasattr(arg, "unit"):
-                error_msg = "a 'unit' attribute without an 'is_equivalent' method"
+                error_msg = ("a 'unit' attribute without an 'is_equivalent' "
+                             "method")
             else:
                 error_msg = "no 'unit' attribute"
 
-            raise TypeError("Argument '{}' to function '{}' has {}. "
-                  "You may want to pass in an astropy Quantity instead."
-                     .format(param_name, func_name, error_msg))
+            raise TypeError(f"Argument '{param_name}' to function '{func_name}'"
+                            f" has {error_msg}. You should pass in an astropy "
+                            "Quantity instead.")
 
     else:
+        error_msg = (f"Argument '{param_name}' to function '{func_name}' must "
+                     "be in units convertible to")
         if len(targets) > 1:
-            raise UnitsError("Argument '{}' to function '{}' must be in units"
-                             " convertible to one of: {}."
-                             .format(param_name, func_name,
-                                     [str(targ) for targ in targets]))
+            targ_names = ", ".join([f"'{str(targ)}'" for targ in targets])
+            raise UnitsError(f"{error_msg} one of: {targ_names}.")
         else:
-            raise UnitsError("Argument '{}' to function '{}' must be in units"
-                             " convertible to '{}'."
-                             .format(param_name, func_name,
-                                     str(targets[0])))
+            raise UnitsError(f"{error_msg} '{str(targets[0])}'.")
 
 
 class QuantityInput:
@@ -89,16 +101,16 @@ class QuantityInput:
         r"""
         A decorator for validating the units of arguments to functions.
 
-        Unit specifications can be provided as keyword arguments to the decorator,
-        or by using function annotation syntax. Arguments to the decorator
-        take precedence over any function annotations present.
+        Unit specifications can be provided as keyword arguments to the
+        decorator, or by using function annotation syntax. Arguments to the
+        decorator take precedence over any function annotations present.
 
         A `~astropy.units.UnitsError` will be raised if the unit attribute of
-        the argument is not equivalent to the unit specified to the decorator
-        or in the annotation.
-        If the argument has no unit attribute, i.e. it is not a Quantity object, a
-        `ValueError` will be raised unless the argument is an annotation. This is to
-        allow non Quantity annotations to pass through.
+        the argument is not equivalent to the unit specified to the decorator or
+        in the annotation. If the argument has no unit attribute, i.e. it is not
+        a Quantity object, a `ValueError` will be raised unless the argument is
+        an annotation. This is to allow non Quantity annotations to pass
+        through.
 
         Where an equivalency is specified in the decorator, the function will be
         executed with that equivalency in force.
@@ -152,9 +164,10 @@ class QuantityInput:
         else:
             return self
 
-    def __init__(self, func=None, **kwargs):
+    def __init__(self, func=None, strict_dimensionless=False, **kwargs):
         self.equivalencies = kwargs.pop('equivalencies', [])
         self.decorator_kwargs = kwargs
+        self.strict_dimensionless = strict_dimensionless
 
     def __call__(self, wrapped_function):
 
@@ -175,7 +188,8 @@ class QuantityInput:
                     continue
 
                 # Catch the (never triggered) case where bind relied on a default value.
-                if param.name not in bound_args.arguments and param.default is not param.empty:
+                if (param.name not in bound_args.arguments
+                        and param.default is not param.empty):
                     bound_args.arguments[param.name] = param.default
 
                 # Get the value of this parameter (argument to new function)
@@ -204,7 +218,8 @@ class QuantityInput:
                 #   were specified in the decorator/annotation, or whether a
                 #   single string (unit or physical type) or a Unit object was
                 #   specified
-                if isinstance(targets, str) or not isinstance(targets, Sequence):
+                if (isinstance(targets, str)
+                        or not isinstance(targets, Sequence)):
                     valid_targets = [targets]
 
                 # Check for None in the supplied list of allowed units and, if
@@ -222,17 +237,21 @@ class QuantityInput:
                 #    are not strings or subclasses of Unit. This is to allow
                 #    non unit related annotations to pass through
                 if is_annotation:
-                    valid_targets = [t for t in valid_targets if isinstance(t, (str, UnitBase))]
+                    valid_targets = [t for t in valid_targets
+                                     if isinstance(t, (str, UnitBase))]
 
                 # Now we loop over the allowed units/physical types and validate
                 #   the value of the argument:
                 _validate_arg_value(param.name, wrapped_function.__name__,
-                                    arg, valid_targets, self.equivalencies)
+                                    arg, valid_targets, self.equivalencies,
+                                    self.strict_dimensionless)
 
             # Call the original function with any equivalencies in force.
             with add_enabled_equivalencies(self.equivalencies):
                 return_ = wrapped_function(*func_args, **func_kwargs)
-            if wrapped_signature.return_annotation not in (inspect.Signature.empty, None):
+
+            valid_empty = (inspect.Signature.empty, None)
+            if wrapped_signature.return_annotation not in valid_empty:
                 return return_.to(wrapped_signature.return_annotation)
             else:
                 return return_