+++ astropy/units/tests/test_quantity_decorator.py
@@ -2,6 +2,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
 import pytest
+import numpy as np
 
 from astropy import units as u
 
@@ -102,7 +103,7 @@ def test_not_quantity(x_input, y_input):
 
     with pytest.raises(TypeError) as e:
         x, y = myfunc_args(1*x_unit, 100)
-    assert str(e.value) == "Argument 'y' to function 'myfunc_args' has no 'unit' attribute. You may want to pass in an astropy Quantity instead."
+    assert str(e.value) == "Argument 'y' to function 'myfunc_args' has no 'unit' attribute. You should pass in an astropy Quantity instead."
 
 
 def test_not_quantity_annotated(x_input, y_input):
@@ -115,7 +116,7 @@ def test_not_quantity_annotated(x_input, y_input):
 
     with pytest.raises(TypeError) as e:
         x, y = myfunc_args(1*x_unit, 100)
-    assert str(e.value) == "Argument 'y' to function 'myfunc_args' has no 'unit' attribute. You may want to pass in an astropy Quantity instead."
+    assert str(e.value) == "Argument 'y' to function 'myfunc_args' has no 'unit' attribute. You should pass in an astropy Quantity instead."
 
 
 def test_kwargs(x_input, y_input):
@@ -180,7 +181,7 @@ def test_kwarg_not_quantity(x_input, y_input):
 
     with pytest.raises(TypeError) as e:
         x, y = myfunc_args(1*x_unit, y=100)
-    assert str(e.value) == "Argument 'y' to function 'myfunc_args' has no 'unit' attribute. You may want to pass in an astropy Quantity instead."
+    assert str(e.value) == "Argument 'y' to function 'myfunc_args' has no 'unit' attribute. You should pass in an astropy Quantity instead."
 
 
 def test_kwarg_default(x_input, y_input):
@@ -283,7 +284,7 @@ def test_no_equivalent():
     with pytest.raises(TypeError) as e:
         x, y = myfunc_args(test_quantity())
 
-        assert str(e.value) == "Argument 'x' to function 'myfunc_args' has a 'unit' attribute without an 'is_equivalent' method. You may want to pass in an astropy Quantity instead."
+        assert str(e.value) == "Argument 'x' to function 'myfunc_args' has a 'unit' attribute without an 'is_equivalent' method. You should pass in an astropy Quantity instead."
 
 
 def test_kwarg_invalid_physical_type():
@@ -376,3 +377,33 @@ def test_args_None_kwarg():
 
     with pytest.raises(TypeError):
         x, y = myfunc_args(None, None)
+
+
+@pytest.mark.parametrize('val', [1., 1, np.arange(10), np.arange(10.)])
+def test_allow_dimensionless_numeric(val):
+    """
+    When dimensionless_unscaled is an allowed unit, numbers and numeric numpy
+    arrays are allowed through
+    """
+
+    @u.quantity_input(velocity=[u.km/u.s, u.dimensionless_unscaled])
+    def myfunc(velocity):
+        return velocity
+
+    assert np.all(myfunc(val) == val)
+
+
+@pytest.mark.parametrize('val', [1., 1, np.arange(10), np.arange(10.)])
+def test_allow_dimensionless_numeric_strict(val):
+    """
+    When dimensionless_unscaled is an allowed unit, but we are being strict,
+    don't allow numbers and numeric numpy arrays through
+    """
+
+    @u.quantity_input(velocity=[u.km/u.s, u.dimensionless_unscaled],
+                      strict_dimensionless=True)
+    def myfunc(velocity):
+        return velocity
+
+    with pytest.raises(TypeError):
+        assert myfunc(val)